requirements.txt
Flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.1.0


